backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.19.0,<0.20.0
plaid-python>=7.2.1
supertokens-python>=0.12.0
python-multipart>=0.0.5
pydantic>=1.8.0,<2.0.0
requests>=2.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-baserow>=0.1.0
aiohttp>=3.8.0
